Add option to generate query for list of products
Modified the `sdm generate-query` command. Added an option to read in a
file with a list of products and generate a query that will return all
of them. The file should contain one product per line without, with no
extensions.

Also updated the default behaviour for generating a query from the
config file. Now if no flags for what to include are provided,
everything will be included.