Improve configuration
Changed the way authentication is extracted from the configuration file.
Now it checks if an option is left as default and skips it if it was.
Also, now the first check is for the environment variables.

Changed the way ROI is extracted from the configuration file.
Same as auth, now the default dummy values are ignored.

Added comments to the configuration file template.