Rewrite product fetch to use separate process
Changed the way the download is handled by `sdm fetch product`. Now it
starts a download not as a thread, but as a separate process. I believe
it will be easier to generalize when handling concurrent downloads in
the future (i.e., download a list of products 2 at a time max).